% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripFrequency.R
\name{as.TransitTime}
\alias{as.TransitTime}
\title{Function to convert HH:MM:SS to seconds after midnight}
\usage{
as.TransitTime(x)
}
\arguments{
\item{x}{character column or single character of times in "HH:MM:SS" format.}
}
\value{
character column or single character in data.table ITime format which displays regular time but is made of integer values
}
\description{
Function to convert HH:MM:SS to seconds after midnight
}
