% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripFrequency.R
\name{calculateFrequencies}
\alias{calculateFrequencies}
\title{Calculate frequencies for gtfs frequencies.txt (frequency in seconds for each trip_id)}
\usage{
calculateFrequencies(gtfs)
}
\arguments{
\item{gtfs}{object from tidytransit read_gtfs()}
}
\value{
data.table conforming to gtfs frequencies.txt format.
}
\description{
Calculate frequencies for gtfs frequencies.txt (frequency in seconds for each trip_id)
}
