% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencyAndHeadway.R
\name{calculateFrequenciesByRoute}
\alias{calculateFrequenciesByRoute}
\title{Calculate frequencies per route and time of day from tidytransit gtfs object}
\usage{
calculateFrequenciesByRoute(gtfs)
}
\arguments{
\item{gtfs}{object from tidytransit read_gtfs()}
}
\value{
data.table containing route, time period, and average headway
}
\description{
Calculate frequencies per route and time of day from tidytransit gtfs object
}
