% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencyAndHeadway.R
\name{calculateFrequenciesByTripAndService}
\alias{calculateFrequenciesByTripAndService}
\title{Calculate frequencies for gtfs frequencies.txt (frequency in seconds for each trip_id-service_id pair). Can take up to 20 seconds or so}
\usage{
calculateFrequenciesByTripAndService(gtfs)
}
\arguments{
\item{gtfs}{object from tidytransit read_gtfs()}
}
\value{
data.table conforming to gtfs frequencies.txt format with times as ITime (integer seconds)
}
\description{
Calculate frequencies for gtfs frequencies.txt (frequency in seconds for each trip_id-service_id pair). Can take up to 20 seconds or so
}
