% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travelTimes.R
\name{calculateMedianTravelTime}
\alias{calculateMedianTravelTime}
\title{Calculate travel time median between two stop_ids on a given route for a time range (default is 5AM-11:30PM)}
\usage{
calculateMedianTravelTime(
  gtfs,
  beginStopID,
  endStopID,
  routeID,
  startHHMMSSRange = "00:05:00",
  endHHMMSSRange = "23:30:00"
)
}
\arguments{
\item{gtfs}{object from gtfsFunctions::formatGTFSObject()}

\item{beginStopID}{stop_id string to begin travel time calculations from}

\item{endStopID}{stop_id string to end travel time calculations at}

\item{routeID}{route_id string to calculate travel times for between stop_ids}

\item{startHHMMSSRange}{beginning HHMMSS string of timeframe to get travel times in (i.e. "15:30:00")}

\item{endHHMMSSRange}{end HHMMSS string of timeframe to get travel times in (i.e. "18:30:00")}
}
\value{
summary of travel times in specified time range
}
\description{
Calculate travel time median between two stop_ids on a given route for a time range (default is 5AM-11:30PM)
}
