% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routesAtStop.R
\name{commaCharacter2List}
\alias{commaCharacter2List}
\title{Function to take string of form '2,4,5,21,56' and convert to a list of form ('2','4','21','56')}
\usage{
commaCharacter2List(row)
}
\arguments{
\item{row}{Leave empty and use as function in lapply}
}
\value{
column of lists
}
\description{
Function to take string of form '2,4,5,21,56' and convert to a list of form ('2','4','21','56')
}
