% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatGtfsFiles.R
\name{formatGTFSObject}
\alias{formatGTFSObject}
\title{Format GTFS zip file into Data.Table Objects with a routes.txt field formattedRouteName that is the route_short_name (or route_id if no short_name)}
\usage{
formatGTFSObject(gtfsZipPath)
}
\arguments{
\item{gtfsZipPath}{character file path to gtfs zip file}
}
\value{
A gtfs/list object with a single data.table for each of the primary gtfs tables (listed above)
}
\description{
This function takes a file path to a gtfs Zip file converts all the primary required files (agency, stops, routes, trips,
stop_times, calendar, calendar_dates, shapes) into data.table objects. URL links are formatted as HTML links for shiny rendering
and dashes are removed in route_id's if the same after dash value occurs more than 10 times (to fix transitfeeds duplication of route_ids),
A new field in the routes.txt file is created (formattedRouteName) that is the route_short_name (or route_id if no short_name). Useful for working with data where route_id doesn't directly correspond to same number that is public-facing
}
