% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapted_tidytransit.R
\name{get_route_frequency}
\alias{get_route_frequency}
\title{Calculate frequencies per route from tidytransit gtfs object (This is addapted from tidytransit frequencies.R). Uses stop_times and trips}
\usage{
get_route_frequency(
  gtfs_obj,
  by_TOD = T,
  startHHMMSS = "06:00:00",
  endHHMMSS = "22:00:00",
  service_ids = c(),
  by_route = T
)
}
\arguments{
\item{gtfs_obj}{object from tidytransit read_gtfs()}

\item{by_TOD}{(optional) TRUE/FALSE indicating if segmenting frequencies by time of day periods or overall. If by_TOD == T don't provide start and end HHMMSS}

\item{startHHMMSS}{(optional) a string HH:MM:SS (default 06:00:00 (6 am))}

\item{endHHMMSS}{(optional) a string HH:MM:SS (default 22:00:00 (10 pm))}

\item{service_ids}{(optional) a string from the calendar dataframe identifying a particular service schedule.}

\item{by_route}{(optional) TRUE/FALSE indicating if segmenting frequencies by route. (defaul TRUE)}
}
\value{
tbl_df headways per route
}
\description{
Calculate frequencies per route from tidytransit gtfs object (This is addapted from tidytransit frequencies.R). Uses stop_times and trips
}
