% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acsDataInServiceArea.R
\name{getACSIntersectionVectors}
\alias{getACSIntersectionVectors}
\title{Return vectors of state and county FIPS codes in service area}
\usage{
getACSIntersectionVectors(gtfs)
}
\arguments{
\item{gtfs}{object from gtfsfunctions::formatGTFSObject()}
}
\value{
list of stateFIPS and countyFIPS intersecting gtfs$stops bbox
}
\description{
Return vectors of state and county FIPS codes in service area
}
