% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acsDataInServiceArea.R
\name{getServiceAreaACS}
\alias{getServiceAreaACS}
\title{Return spatial polygon acs data within transit service area (defined by stops.txt bbox)}
\usage{
getServiceAreaACS(
  gtfs,
  variables,
  geography = "tract",
  year = 2016,
  survey = "acs5"
)
}
\arguments{
\item{gtfs}{object from gtfsfunctions::formatGTFSObject()}

\item{variables}{Character string or vector of character strings of variable IDs. tidycensus automatically returns the estimate and the margin of error associated with the variable.}

\item{geography}{String string for tidycensus:;get_acs() (i.e. "tract", or "county") default is tract}

\item{year}{int  The year, or endyear, of the ACS sample. 5-year ACS data is available from 2009 through 2019. 1-year ACS data is available from 2005 through 2019. Defaults to 2016}

\item{survey}{string for the ACS contains one-year, three-year, and five-year surveys expressed as "acs1", "acs3", and "acs5". The default selection is "acs5."}
}
\value{
sf polygon of requested acs data in transit area
}
\description{
Return spatial polygon acs data within transit service area (defined by stops.txt bbox)
}
