% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs2sf.R
\name{gtfs2RouteLines}
\alias{gtfs2RouteLines}
\title{Convert GTFS Route, Trip, and Shape File to SF Linestring Object}
\usage{
gtfs2RouteLines(routesDf_GTFS, tripsDf_GTFS, shapesDf_GTFS)
}
\arguments{
\item{routesDf_GTFS}{routes.txt gtfs data.table file}

\item{tripsDf_GTFS}{trips.txt gtfs data.table file}

\item{shapesDf_GTFS}{shapes.txt gtfs data.table file}
}
\value{
A sf object of route geometry linestrings
}
\description{
This function loads and joins gtfs route.txt, trip.txt, and shape.txt files
identifies route colors (if present) and returns an sf object
of linestring features for each shape. Each output record
contains both a shape_id and route_id for easy filtering.
}
