% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routesAtStop.R
\name{namedList_RoutesAtStops}
\alias{namedList_RoutesAtStops}
\title{Faster function to simplify gtfsFunctions::routeIDAtStops() to include only one record per stop with a column containing a list of all routes at stop.}
\usage{
namedList_RoutesAtStops(gtfsObject)
}
\arguments{
\item{gtfsObject}{stops table from gtfsFunctions::routeIDAtStops() or stops.txt with one row per route-stop combo. MUST have "route_id" column}
}
\value{
Names list of all routes at stop where stop is name and routes are values
}
\description{
Faster function to simplify gtfsFunctions::routeIDAtStops() to include only one record per stop with a column containing a list of all routes at stop.
}
