% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routesAtStop.R
\name{simpleRoutesAtStops}
\alias{simpleRoutesAtStops}
\title{Function To Simplify gtfsFunctions::routeIDAtStops() to include only one record per stop with a column containing a list of all routes at stop.}
\usage{
simpleRoutesAtStops(gtfsObject)
}
\arguments{
\item{gtfsObject}{stops table from gtfsFunctions::routeIDAtStops() or stops.txt with one row per route-stop combo. MUST have "route_id" column}
}
\value{
data.table with one record per stop and new field 'routesAtStop' that is a list of all routes at stop
}
\description{
Function To Simplify gtfsFunctions::routeIDAtStops() to include only one record per stop with a column containing a list of all routes at stop.
}
