% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routesAtStop.R
\name{simplifyRoutesAtStops}
\alias{simplifyRoutesAtStops}
\title{Function To Simplify gtfsFunctions::routesAtStops() to include only one record per stop with a column containing a list of all routes at stop.}
\usage{
simplifyRoutesAtStops(stopsTable)
}
\arguments{
\item{stopsTable}{stops table from gtfsFunctions::routesAtStops() or stops.txt with one row per route-stop combo. MUST have "route_id" column}

\item{route_ColumnName}{route_id column name character}
}
\value{
data.table with one record per stop and new field 'routesAtStop' that is a list of all routes at stop
}
\description{
Function To Simplify gtfsFunctions::routesAtStops() to include only one record per stop with a column containing a list of all routes at stop.
}
