% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeFormatting.R
\name{todTable}
\alias{todTable}
\title{Function to create time of day (TOD) data.table}
\usage{
todTable(
  earlyStart = 14400,
  AMPeakStart = 21600,
  MiddayStart = 32400,
  PMPeakStart = 54000,
  EveningStart = 66600,
  NightStart = 75600,
  NightEnd = 86400
)
}
\arguments{
\item{earlyStart}{integer seconds past midnight at which the early period starts}

\item{AMPeakStart}{integer seconds past midnight at which the AMPeak period starts}

\item{MiddayStart}{integer seconds past midnight at which the midday period starts}

\item{PMPeakStart}{integer seconds past midnight at which the PMPeak period starts}

\item{EveningStart}{integer seconds past midnight at which the evening period starts}

\item{NightStart}{integer seconds past midnight at which the night period starts}

\item{NightEnd}{integer seconds past midnight at which the night period starts}
}
\value{
character column or single character in character "HH:MM:SS" format
}
\description{
Function to create time of day (TOD) data.table
}
