% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeFormatting.R
\name{transitTime2HHMMSS}
\alias{transitTime2HHMMSS}
\title{Function to convert seconds after midnight to HH:MM:SS}
\usage{
transitTime2HHMMSS(x)
}
\arguments{
\item{x}{character column or single character of times in integer or ITime format.}
}
\value{
character column or single character in character "HH:MM:SS" format
}
\description{
Function to convert seconds after midnight to HH:MM:SS
}
