% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueRoutesInHexAreaByCounty.R
\name{uniqueRoutesInHexTessalation}
\alias{uniqueRoutesInHexTessalation}
\title{Return spatial polygon of hexagonal features with a count of the number of unique routes with a stop inside the hex area's contained in the specified counties}
\usage{
uniqueRoutesInHexTessalation(
  gtfs,
  stateAbbrev = NULL,
  counties = NULL,
  tidyCensusAPIKey = NULL,
  hexSize = 0.01
)
}
\arguments{
\item{gtfs}{object from gtfsfunctions::formatGTFSObject()}

\item{stateAbbrev}{Character string or vector of character strings of state abbreviations.}

\item{counties}{Character string or vector of character strings of desired counties in state.}

\item{tidyCensusAPIKey}{String for tidy_census::get_acs() api key}

\item{hexSize}{number specifying hex size (smaller value creates more smaller sized hexes but takes longer to load). Default = 0.01}
}
\value{
returns an sf data frame with 4 columns (hex_id, routes, uniqueRouteCount, geometry), where routes contains a character vector of routes in the hex and uniqueRouteCount counts the number of routes in each hex
}
\description{
Return spatial polygon of hexagonal features with a count of the number of unique routes with a stop inside the hex area's contained in the specified counties
}
