% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acsDataInServiceArea.R,
%   R/uniqueRoutesInHexAreaByCounty.R
\name{getServiceAreaACS}
\alias{getServiceAreaACS}
\title{Return spatial polygon acs data within transit service area (defined by stops.txt bbox)}
\usage{
getServiceAreaACS(
  gtfs,
  stateAbbrev = NULL,
  counties = NULL,
  tidyCensusAPIKey = NULL,
  hexSize = 0.01
)

getServiceAreaACS(
  gtfs,
  stateAbbrev = NULL,
  counties = NULL,
  tidyCensusAPIKey = NULL,
  hexSize = 0.01
)
}
\arguments{
\item{gtfs}{object from gtfsfunctions::formatGTFSObject()}

\item{stateAbbrev}{Character string or vector of character strings of state abbreviations.}

\item{counties}{Character string or vector of character strings of desired counties in state.}

\item{tidyCensusAPIKey}{String for tidy_census::get_acs() api key}

\item{hexSize}{number specifying hex size (smaller value creates more smaller sized hexes but takes longer to load). Default = 0.01}

\item{variables}{Character string or vector of character strings of variable IDs. tidycensus automatically returns the estimate and the margin of error associated with the variable.}

\item{geography}{String string for tidycensus:;get_acs() (i.e. "tract", or "county") default is tract}

\item{year}{int  The year, or endyear, of the ACS sample. 5-year ACS data is available from 2009 through 2019. 1-year ACS data is available from 2005 through 2019. Defaults to 2016}

\item{survey}{string for the ACS contains one-year, three-year, and five-year surveys expressed as "acs1", "acs3", and "acs5". The default selection is "acs5."}
}
\value{
sf polygon of requested acs data in transit area

returns an sf data frame with 4 columns (hex_id, routes, uniqueRouteCount, geometry), where routes contains a character vector of routes in the hex and uniqueRouteCount counts the number of routes in each hex
}
\description{
Return spatial polygon acs data within transit service area (defined by stops.txt bbox)

Return spatial polygon of hexagonal features with a count of the number of unique routes with a stop inside the hex area's contained in the specified counties
}
